Move extra specification into uv-configuration
