+++ crates/uv-configuration/src/lib.rs
@@ -3,6 +3,7 @@ pub use build_options::*;
 pub use concurrency::*;
 pub use config_settings::*;
 pub use constraints::*;
+pub use extras::*;
 pub use name_specifiers::*;
 pub use overrides::*;
 pub use package_options::*;
@@ -14,6 +15,7 @@ mod build_options;
 mod concurrency;
 mod config_settings;
 mod constraints;
+mod extras;
 mod name_specifiers;
 mod overrides;
 mod package_options;
